// src/app/support-management/models/truck-details.ts
export class TruckDetails {
    vehiclePlateNumber: string;
    subType: string;
    truckStatus: string;
    fastagNumber: string;
    fastagLinkingStatus: string;
    fastagBalance: number;
    fuelCardNumber: string;
    fuelCardLinkingStatus: string;
    fuelCardBalance: number;
    fuelCardCustomerId: string;
    fuelAutoRechargeStatus: string;
    tollAutoRechargeStatus: string;
}